Remove additional classes
